Better levels

+++ src/edu/mhs/wombat/game/HUDSystem.java
@@ -92,7 +92,7 @@ public class HUDSystem {
 
 	public void render(StateBasedGame game, GameStatus gs, Graphics g) {
 		g.setFont(ResourceManager.getFont("font20"));
-		g.drawString("level : " + gs.levelManager.difficulty, 10, 70);
+		g.drawString("level : " + (int)gs.levelManager.difficulty, 10, 70);
 		g.drawString("score : " + gs.scores.getScore(), 10, 90);
 		g.drawString("multiplier : " + gs.scores.getMultiplier(), 10, 110);
 		g.drawString("weapon : " + gs.player.weps.getName(), 10, 130);
